HUUUUUGE: Moving stuff to view model, but statistics will crash on sort
